# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vartalaap API",
    "description": "Voice bot platform for local Indian businesses",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/knowledge": {
      "get": {
        "tags": [
          "Knowledge",
          "Knowledge"
        ],
        "summary": "List Knowledge Items",
        "description": "List knowledge items with optional filters.\n\nSecurity: business_id is required to prevent cross-tenant data access.",
        "operationId": "list_knowledge_items_api_knowledge_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Required for tenant isolation",
              "title": "Business Id"
            },
            "description": "Required for tenant isolation"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeItemResponse"
                  },
                  "title": "Response List Knowledge Items Api Knowledge Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Knowledge",
          "Knowledge"
        ],
        "summary": "Create Knowledge Item",
        "description": "Create a new knowledge item.",
        "operationId": "create_knowledge_item_api_knowledge_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/{item_id}": {
      "get": {
        "tags": [
          "Knowledge",
          "Knowledge"
        ],
        "summary": "Get Knowledge Item",
        "description": "Get a knowledge item by ID.",
        "operationId": "get_knowledge_item_api_knowledge__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Knowledge",
          "Knowledge"
        ],
        "summary": "Update Knowledge Item",
        "description": "Update a knowledge item (partial update).",
        "operationId": "update_knowledge_item_api_knowledge__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Knowledge",
          "Knowledge"
        ],
        "summary": "Delete Knowledge Item",
        "description": "Delete a knowledge item.",
        "operationId": "delete_knowledge_item_api_knowledge__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/search": {
      "post": {
        "tags": [
          "Knowledge",
          "Knowledge"
        ],
        "summary": "Search Knowledge",
        "description": "Search knowledge items using text matching.\n\nTODO: Integrate with ChromaDB for vector similarity search.\nCurrent implementation uses simple text matching.",
        "operationId": "search_knowledge_api_knowledge_search_post",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "title": "Query"
            }
          },
          {
            "name": "business_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Business Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeSearchResult"
                  },
                  "title": "Response Search Knowledge Api Knowledge Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeCategory": {
        "type": "string",
        "enum": [
          "menu_item",
          "faq",
          "policy",
          "announcement"
        ],
        "title": "KnowledgeCategory",
        "description": "Category of knowledge item."
      },
      "KnowledgeItemCreate": {
        "properties": {
          "business_id": {
            "type": "string",
            "title": "Business Id"
          },
          "category": {
            "$ref": "#/components/schemas/KnowledgeCategory"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "title_hindi": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Hindi"
          },
          "content": {
            "type": "string",
            "maxLength": 2000,
            "title": "Content"
          },
          "content_hindi": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hindi"
          },
          "metadata_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "priority": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "business_id",
          "category",
          "title",
          "content"
        ],
        "title": "KnowledgeItemCreate",
        "description": "Schema for creating a knowledge item."
      },
      "KnowledgeItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "business_id": {
            "type": "string",
            "title": "Business Id"
          },
          "category": {
            "$ref": "#/components/schemas/KnowledgeCategory"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "title_hindi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Hindi"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "content_hindi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hindi"
          },
          "metadata_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "embedding_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "business_id",
          "category",
          "title",
          "title_hindi",
          "content",
          "content_hindi",
          "metadata_json",
          "is_active",
          "priority",
          "embedding_id",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeItemResponse",
        "description": "Response schema for knowledge items."
      },
      "KnowledgeItemUpdate": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "title_hindi": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Hindi"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "content_hindi": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hindi"
          },
          "metadata_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Json"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "type": "object",
        "title": "KnowledgeItemUpdate",
        "description": "Schema for updating a knowledge item (all fields optional)."
      },
      "KnowledgeSearchResult": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/KnowledgeItemResponse"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "item",
          "score"
        ],
        "title": "KnowledgeSearchResult",
        "description": "Search result with similarity score."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}